fix: upcast logits to float32 before log_softmax for KL divergence
Low-precision dtypes (bfloat16/float16) cause log_softmax to produce
-inf for low-probability tokens. When abliteration shifts which tokens
underflow, kl_div returns NaN/inf. Upcasting to float32 matches the
existing pattern used for residual vectors.

Also reverts the max_weight range change from 2e6cfc6 since the search
ranges were not the actual problem.

+++ src/heretic/model.py
@@ -717,7 +717,9 @@ class Model:
         logits = cast(tuple[FloatTensor], outputs.scores)[0]
 
         # The returned tensor has shape (prompt, token).
-        return F.log_softmax(logits, dim=-1)
+        # Upcast to float32 before log_softmax to avoid -inf from underflow
+        # in low-precision dtypes (FP8, float16, bfloat16).
+        return F.log_softmax(logits.float(), dim=-1)
 
     def get_logprobs_batched(self, prompts: list[Prompt]) -> Tensor:
         logprobs = []
